Basic working code ready with filter
